Change return type from BigInteger to Long

+++ src/main/java/com/mev/films/mappers/interfaces/UserMapper.java
@@ -2,13 +2,12 @@ package com.mev.films.mappers.interfaces;
 
 import com.mev.films.model.UserDTO;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface UserMapper {
     List<UserDTO> selectUsers();
     UserDTO selectUserByName(String userName);
-    BigInteger insertUser(UserDTO userDTO);
-    BigInteger updateUser(UserDTO userDTO);
-    BigInteger removeUserByName(String userName);
+    Long insertUser(UserDTO userDTO);
+    Long updateUser(UserDTO userDTO);
+    Long removeUserByName(String userName);
 }
